refactor(price): refactorize price estimation
